Implement adding new items to menu (backend only)
It is now possible to add new items to the menu, on the backend only (for now).
Also, for now, the model for menu items requires "hardcoded" names for category and ingredients fields.
Made overall changes to the menuItem db model (no more description field, added fields: category, ingredients, timesSold).