Format files and improve error handling
Aside from just formatting all files in the project folder for the sake of organization and clean code,
slight changes have also been made to error handling.
Now, when wanting to throw an error, the next function is instantly called on the custom error instance.
This is done to preserve clarity of this part of the code.